Let me make a closing commit for the day so the daily commits are not entirely made of journal entries.
I think today went splendidly. I did what I wanted to do in its entirety. And in the days that are to follow I will thoroughly get this architecture to work as it should. I will bend it to my will and draw out its power.

Besides this test, I really should look into the maze test as well. There is just no point in rushing this.

Plasticity is one of the biggest architectural innovations in the deep learning era. All the work that is holding me back here will come back to reinforce me in the battles to come.

But for now, it is time for a much needed rest.