I lose. I cannot think of a way to implement wholesale PRONG for the `mi_hebb` layer without necessarily doing it all by hand. What I will do instead is just use PRONG for the matrix multiplies. I suppose for the biases I could always just do it in the optimization stage. That would be the best choice. There I could do PRONG directly on the adjoints.
Damn, why isn't programming easier? It is just so hard unless I take the easy paths.

Let me replace the usual matrix multiplies with naturalized ones...after the breakfast.

I'll make finishing this layer my goal for the day. Well, it won't be too hard.

Damn, it would be so great if I could solve all my problems by just moving a bunch of pointers around. I spent over half an hour thinking along those lines. And no it would not be possible. I can't let the gradients spill because it would change the functionality of the net every time the covariance matrices are updated.

Operating on gradients of the bias terms directly is a good route to take. I think at some point I will put `n` back in and to make it work, I'll use NG on it too directly.

When I get back I'll really focus on making this work as it should.