"1:30pm.
```
The learning rate is 2 ** -14
At iteration 9 the cost is 50.0108589172363
At iteration 19 the cost is 50.0321949005127
At iteration 29 the cost is 49.9644626617432
At iteration 39 the cost is 49.7141738891602
At iteration 49 the cost is 49.7676620483398
At iteration 59 the cost is 49.5293338775635
At iteration 69 the cost is 49.7837940216064
At iteration 79 the cost is 49.4830291748047
At iteration 89 the cost is 49.0657279968262
At iteration 99 the cost is 49.1555801391602
At iteration 109 the cost is 48.6704196929932
At iteration 119 the cost is 48.8506973266602
At iteration 129 the cost is 48.778381729126
At iteration 139 the cost is 48.2755878448486
At iteration 149 the cost is 48.1645893096924
At iteration 159 the cost is 48.2874000549316
At iteration 169 the cost is 47.9468910217285
At iteration 179 the cost is 47.7391181945801
At iteration 189 the cost is 47.2731716156006
At iteration 199 the cost is 47.2969192504883
At iteration 209 the cost is 47.2216136932373
At iteration 219 the cost is 46.7442081451416
At iteration 229 the cost is 46.9328147888184
At iteration 239 the cost is 46.9551708221436
At iteration 249 the cost is 46.8961750030518
At iteration 259 the cost is 46.415225982666
At iteration 269 the cost is 46.4221141815186
At iteration 279 the cost is 46.1211078643799
At iteration 289 the cost is 45.5560878753662
At iteration 299 the cost is 45.6908603668213
At iteration 309 the cost is 45.5671073913574
At iteration 319 the cost is 45.0989593505859
At iteration 329 the cost is 45.3176094055176
At iteration 339 the cost is 44.9055183410645
At iteration 349 the cost is 45.0720733642578
At iteration 359 the cost is 44.8052555084229
At iteration 369 the cost is 44.0872402191162
At iteration 379 the cost is 44.1949043273926
At iteration 389 the cost is 43.4550216674805
At iteration 399 the cost is 43.0241130828857
At iteration 409 the cost is 43.2289356231689
At iteration 419 the cost is 42.857201385498
At iteration 429 the cost is 42.1050392150879
At iteration 439 the cost is 42.2785774230957
At iteration 449 the cost is 41.6907997131348
At iteration 459 the cost is 41.2549327850342
At iteration 469 the cost is 40.5930530548096
At iteration 479 the cost is 41.0526351928711
At iteration 489 the cost is 40.6866115570068
At iteration 499 the cost is 40.2835918426514
At iteration 509 the cost is 40.1615383148193
At iteration 519 the cost is 40.0999759674072
At iteration 529 the cost is 39.6593456268311
At iteration 539 the cost is 40.8000747680664
At iteration 549 the cost is 40.28088722229
At iteration 559 the cost is 42.2477600097656
At iteration 569 the cost is 40.5876457214356
At iteration 579 the cost is 40.8975650787354
At iteration 589 the cost is 38.3219718933105
At iteration 599 the cost is 41.7053577423096
At iteration 609 the cost is 39.639465713501

Unhandled Exception: ManagedCuda.CudaBlas.CudaBlasException: ExecutionFailed: An access to GPU memory space failed.
```

No it won't work. It just so unstable even when I've ratched down the learning rate.

1:45pm. Hmmm...I wonder what the true story is here? I mean, would't the cost blow up a little before the thing goes haywire? What if I set the epsilon to something really high? to check it out. Also I'll set the printing for each iteration to 1.

Before I conclude that my methods of passing gradients into W is wrong, I need to make sure that is in fact the problem."