2:10pm. https://math.stackexchange.com/questions/2910722/is-there-a-way-to-distribute-a-h-x-so-it-becomes-a-y
This is so annoying.

2:30pm. I am stuck. I am not sure what to do here. But even though PRONG is blowing up, should this really be happening? Could it be because I am doing online updates? Let me set n so the thing is done only 32 per iteration...

```
At iteration 108 the cost is 46.400074005127

Unhandled Exception: ManagedCuda.CudaException: ErrorAssert: A device-side assert triggered during kernel execution. The context cannot be used anymore, and must be destroyed. All existing device memory allocations from this context are invalid and must be reconstructed if the program is to continue using CUDA.
```

Hmmm, here it even didn't blow up. Why isn't this a standard matrix inversion failure?

What is going on here? Let me try it on a really small example. I'll print out the matrices as the thing runs...