I really was going to do work on this, but every time I thought about doing it my thoughts got sidetracked into design.
This issue of how to naturalize the alpha is not something I expected. Given that n was blowing up when I was first testing prong my mind was always open to the possibility that my briliant idea of letting the gradients through the timesteps in a naturalized fashion might not turn out well.

But that does not mean I thought through the implications of that. And I am surprised to find that I cannot think of even a single way of how to reproject the alpha on the backward step.

Not without breaking the entire construction and hacking it to death in order to get what I want.

Hence I considered another possibility. Given how long these sequences are - 200 steps it is actually a perfectly viable choice for RNNs to do the reprojection on the optimization step instead. This is even with online learning.

In this case I would not have to mess with figuring out what alpha wants.

I could still update the covariance matrices on the backward steps. I could also do buffered updates of the covariance matrices to speed things up even further.

...Well, I'll save covariance buffering for later.

What I need to do here is finally put some thought into the optimization pass. I'll get rid how I was doing it up to now and give it my all into making the new thing work.

This will also open the door to dealing with those biases. And `n` should I chose to will be an easy thing to naturalize.

I think I have an implementation of how to do the reprojection in the past version of the library, but that was such shoody, hasty work. I really just only made that to get it to work in a hurry.

I am going to do it much better this time.

As for the `naturalize_matmultb` I am considering removing it, but it definitely has its uses in feedforward networks.

Right now each hand of poker has maybe a dozen hands, but as I will want to maximize winrates for tournaments rather than EV on individual hands, I am going to have to move to longer sequences eventually, so that will help amortize the costs of PRONG.

Let me call it a day here. Given that it is 5:30pm I need a break at this point. I am not going to be starting new things here. Laziness (after a hard day's work) is the cure for stress after all.