"11am. Yeah, that fixed it. Ok.
Let me try running it again.

```
The learning rate is 2 ** -12
At iteration 9 the cost is 49.9145195007324
At iteration 19 the cost is 49.8993064880371
At iteration 29 the cost is 50.0725723266602
At iteration 39 the cost is 49.8159660339355
At iteration 49 the cost is 49.966686630249
At iteration 59 the cost is 49.6550430297852
At iteration 69 the cost is 49.733296585083
At iteration 79 the cost is 49.0862808227539
At iteration 89 the cost is 49.3674160003662
At iteration 99 the cost is 48.7599636077881
At iteration 109 the cost is 49.0260299682617
```

This might be too high of a learning rate. Let me try lowering it a tad. But at least it is not blowing up anymore.

...No actually it might be too low instead.

11:10am. This is not working no matter what learning rates I try. It is rather strange - it barely moves from 50 no matter what I set.

11:20am. No I can't get it to work. I am not sure why. Maybe the updates for `alpha` are incorrect. I've figured out how that one should be done just now...no wait. I can't do that kind of thing as it would be too expensive.

There is something I need to make sure of.

I am going to have to go to the char RNN and try the `mi_prong` redone similarly to this one there. That will tell me whether the thing is correct or not. I need that sort of sanity check. How complicated.

Well, these difficulties are only to be expected. I'll think of what will happen after I get the char RNN to work. Let me take a break here."