"12:50pm. Ok, that was easy. Let me get it to compile and I'll run it.
```
The learning rate is 2 ** -9.999356
Starting timing for: iteration 0
CuSolve function failed!
The leading minor of order 23 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:271: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.

Unhandled Exception: ManagedCuda.CudaException: ErrorAssert: A device-side assert triggered during kernel execution. The context cannot be used anymore, and must be destroyed. All existing device memory allocations from this context are invalid and must be reconstructed if the program is to continue using CUDA.
```

Ah, this one fails immediately. Let me try raising the learning rate.

```
The learning rate is 2 ** -14.00083
Starting timing for: iteration 0
CuSolve function failed!
The leading minor of order 23 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:268: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

I get this no matter what I set the learning rate to.

One more time. This time -16. Though it is 100% certain that I made some kind of mistake.

1pm.

```
    inl with_zero_adjoints x s =
        inl out = Struct.map (inl x -> {x with adjoint=s.CudaTensor.zero_like self}) x
        {
        out
        bck=met _ ->
            //inl x = Struct.map (inl {adjoint} -> adjoint) x
            //inl out = Struct.map (inl {adjoint} -> adjoint) out
            //s.CudaFun.map {out=x; map=inl a,b -> Struct.map2 (+) a b} (x,out)
            ()
        }
```

Even disabling the gradient flow to the biases does not help any. I thought it might be the biases, but I guess I am wrong.

```
The learning rate is 2 ** -9.000094
Starting timing for: iteration 0
The time was 00:00:07.4670590 for: iteration 0
Training: 4.1470649976511
The learning rate is 2 ** -9.000094
Starting timing for: iteration 1
The time was 00:00:07.4886498 for: iteration 1
Training: 3.80105876122252
The learning rate is 2 ** -9.000094
Starting timing for: iteration 2
The time was 00:00:07.4898932 for: iteration 2
Training: 3.78150919707824
The learning rate is 2 ** -9.000094
Starting timing for: iteration 3
The time was 00:00:07.4600679 for: iteration 3
Training: 3.7913092079295
The learning rate is 2 ** -9.000094
Starting timing for: iteration 4
The time was 00:00:07.4957164 for: iteration 4
Training: 3.77556363236234
```

Hmmm...this is without the updates to the covariance matrices.

The architecture is working incredibly poorly.

What if I used `mi_hebb`?"