"```                             fwd=inl {input state bias={si s i c}} -> si * state * input + s * state + i * input + c |> tanh_fwd                             bck=inl {in={in with input state} out} ->                                 inl out = tanh_bck out                                 {                                 input = si * state + i                                 state = si * input + s ```
It should be like this.

```
    inl generalized_mi =
        activation {
            fwd=inl {input state bias={si s i c}} -> si * state * input + s * state + i * input + c |> tanh_fwd
            bck=inl {in={in with input state} out} ->
                inl out = tanh_bck out
                {
                input = si * state + i
                state = si * input + s
                bias = { si = input*state; i = input; s = state; c = one }
                } |> Struct.map ((*) out)
            }
```

I'll factor it out. I have the incorrect version copy pasted everywhere.

1:35pm.

```
The learning rate is 2 ** -9.999356
Starting timing for: iteration 0
The time was 00:00:04.8125252 for: iteration 0
Training: 3.47800726386006
The learning rate is 2 ** -9.999356
Starting timing for: iteration 1
The time was 00:00:04.9995620 for: iteration 1
Training: 3.07648786478197
The learning rate is 2 ** -9.999356
Starting timing for: iteration 2
The time was 00:00:04.7332229 for: iteration 2
Training: 2.83204128655352
The learning rate is 2 ** -9.999356
Starting timing for: iteration 3
The time was 00:00:04.6154025 for: iteration 3
Training: 2.66817157717785
The learning rate is 2 ** -9.999356
Starting timing for: iteration 4
The time was 00:00:04.7505582 for: iteration 4
Training: 2.56006838282337
```

Yeah, now it works. Though this also means the the mi_hebb test is wrong. I am going to have to repeat it.

Let me increase the lr just so I can make sure that it gets to 1.77.

```
The learning rate is 2 ** -6.999909
Starting timing for: iteration 0
The time was 00:00:04.8379874 for: iteration 0
Training: 2.86693894662612
The learning rate is 2 ** -6.999909
Starting timing for: iteration 1
The time was 00:00:04.6838130 for: iteration 1
Training: 2.29211503016535
The learning rate is 2 ** -6.999909
Starting timing for: iteration 2
The time was 00:00:04.7595384 for: iteration 2
Training: 2.08988359153901
The learning rate is 2 ** -6.999909
Starting timing for: iteration 3
The time was 00:00:04.6757305 for: iteration 3
Training: 1.95038836879025
The learning rate is 2 ** -6.999909
Starting timing for: iteration 4
The time was 00:00:05.0212842 for: iteration 4
Training: 1.84368179287214
```

Close enough.

Let me try the `mi_prong_alt` again."