I am up. I had some time to think about this. I do not think I am heading in the right direction with differentiable plasticity. It is too memory hungry and it bothers me how adding a second layer causes the performance to fall off a cliff. It is the kind of algorithm that would spike in value if a way to unbias the credit assignment with very short truncated sequences would ever get found, but now it is of limited use.
I am thinking that I would be better served with a SNAIL like architecture. I should the the mi_alt and instead of passing it the state, I should pass it the output of the lower layer instead. And for each layer stage it by 2x.

But I do not feel like working on a whole new architecture right now. What I should do is that that mi_hebb_prong and test it on poker. Also add a second layer and see if it breaks down. If it is only on par with `mi`, then I am wasting my time with this.

Let me chill a little and then I will start.