``` The learning_rate is 2 ** -9 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 290 and 710 out of 1000. The time was 00:00:07.4913422 for: iteration 0 Starting timing for: iteration 1 Winrate is 577 and 423 out of 1000. The time was 00:00:08.5942470 for: iteration 1 Starting timing for: iteration 2 Winrate is 630 and 370 out of 1000. The time was 00:00:06.5365468 for: iteration 2 Starting timing for: iteration 3 Winrate is 716 and 284 out of 1000. The time was 00:00:06.7729329 for: iteration 3 Starting timing for: iteration 4 Winrate is 709 and 291 out of 1000. The time was 00:00:05.8691777 for: iteration 4 Starting timing for: iteration 5 Winrate is 717 and 283 out of 1000. The time was 00:00:06.2018049 for: iteration 5 Starting timing for: iteration 6 Winrate is 705 and 295 out of 1000. The time was 00:00:06.2465988 for: iteration 6 Starting timing for: iteration 7 Winrate is 690 and 310 out of 1000. The time was 00:00:05.6250033 for: iteration 7 Starting timing for: iteration 8 Winrate is 717 and 283 out of 1000. The time was 00:00:05.3211752 for: iteration 8 Starting timing for: iteration 9 Winrate is 718 and 282 out of 1000. The time was 00:00:05.2881929 for: iteration 9 Starting timing for: iteration 10 Winrate is 730 and 270 out of 1000. The time was 00:00:05.4639490 for: iteration 10 Starting timing for: iteration 11 Winrate is 723 and 277 out of 1000. The time was 00:00:05.6230759 for: iteration 11 Starting timing for: iteration 12 Winrate is 717 and 283 out of 1000. The time was 00:00:05.1258133 for: iteration 12 Starting timing for: iteration 13 Winrate is 731 and 269 out of 1000. The time was 00:00:05.4343346 for: iteration 13 Starting timing for: iteration 14 Winrate is 726 and 274 out of 1000. The time was 00:00:05.1370487 for: iteration 14 ```
This is the `mi_alt` without prong. Not bad at all. Let me try the mi_hebb. Also, I had to strip prong from the actor and the critic. It seems that using PRONG only in some places causes issues. Makes sense I guess.

I made a little question on the RL sub regarding setting of the learning rates for the critic.

...Oh wow, what are these results. It has to be a fluke.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 234 and 766 out of 1000.
The time was 00:00:10.5152404 for: iteration 0
Starting timing for: iteration 1
Winrate is 539 and 461 out of 1000.
The time was 00:00:16.6754862 for: iteration 1
Starting timing for: iteration 2
Winrate is 596 and 404 out of 1000.
The time was 00:00:17.2521805 for: iteration 2
Starting timing for: iteration 3
Winrate is 681 and 319 out of 1000.
The time was 00:00:17.8291002 for: iteration 3
Starting timing for: iteration 4
Winrate is 740 and 260 out of 1000.
The time was 00:00:17.6658306 for: iteration 4
Starting timing for: iteration 5
Winrate is 731 and 269 out of 1000.
The time was 00:00:18.0828048 for: iteration 5
Starting timing for: iteration 6
Winrate is 720 and 280 out of 1000.
The time was 00:00:17.8172582 for: iteration 6
Starting timing for: iteration 7
Winrate is 702 and 298 out of 1000.
The time was 00:00:17.6543164 for: iteration 7
Starting timing for: iteration 8
Winrate is 730 and 270 out of 1000.
The time was 00:00:16.8928095 for: iteration 8
Starting timing for: iteration 9
Winrate is 726 and 274 out of 1000.
The time was 00:00:16.7419653 for: iteration 9
Starting timing for: iteration 10
Winrate is 720 and 280 out of 1000.
The time was 00:00:17.1246568 for: iteration 10
Starting timing for: iteration 11
Winrate is 697 and 303 out of 1000.
The time was 00:00:17.6669813 for: iteration 11
Starting timing for: iteration 12
Winrate is 711 and 289 out of 1000.
The time was 00:00:16.8268157 for: iteration 12
Starting timing for: iteration 13
Winrate is 717 and 283 out of 1000.
The time was 00:00:16.3944908 for: iteration 13
Starting timing for: iteration 14
Winrate is 696 and 304 out of 1000.
The time was 00:00:15.6770642 for: iteration 14
```

This is mi_hebb with a n of 0.04.

Let me try `mi_alt` again with a lowered learning rate. After that I will try stacking two layers and seeing what happens.