"12:25pm. Done with breakfast and Tonegawa. I nearly forgot that I had that choco milk saved away from two days ago.
Let me collect my thoughts. It seems that regardless of what I choose, PRONG will be a vital ingredient. Poker is just not doable without it. That thing that I thought was a fluke in fact was. Without PRONG, the `mi_hebb'_prong` net is only slightly less stable than a regular one. It might very well be the case that the net is not even using plasticity.

This slight decrease in stability is a bad thing in my eyes. In RL, the winrate goes straight into profitability. Hence I cannot compromise on stability under any circumstance. A net that cannot learn to play optimally against a static player 100% of the time is worthless to me.

1pm. I am still thinking about it. I said I would do the maze game, but that thing takes 250 steps per episode and looking at the paper goes for 100ks of episodes. I simply do not have the time to iterate on that. I might be able to close the gap a little by doing it in parallel, but regardless, I think I should just make a better test.

1:20pm. I've been thinking about poker. It seems that the time to start doing it for real is slowly approaching.

Why do I need metalearning? A feedforward player could learn to play unexploitably, but if I want to exploit weak players I need an agent that can adapt in real time. Hence my great need for nets with memory.

Maybe I could let a RNN agent play versus against a feedforward and that will tell me where things stand, but I need to take it a step further.

Since NN players will eventually reach the Nash equilibruim, what I should do is take a trained player and give it exploitable tendencies. In real life, some players fold too much, some raise too much and some call too much. Those kinds of players could be arranged. There is no point in writing them by hand. Rather I should just train and hack.

I'll increase the stack size to something like 50 and allow a wider range of possible bets. Not literally everything, but quarter pot, half pot, pot, all in and so on should be fine. I'll leave absolute flexibility for real life battles. Right now I want fast iteration. So the game does not end up taking forever, like in standard HU SNGs I'll increase the blinds every once in a while. Every 10 hands should be fine.

1:40pm. How goes the test with 2 layers? I am still thinking about this.

It seems to be doing quite well. I thought I might need to tune the learning rates, but it seems -10 is fine even for two layers.

1:45pm. Let me resume the thread of thought. When the two layer test for plastic RNNs is done I will be done with it for the time being.

Right now I am thinking of how to make tendencies for feedforward nets. I think I know how. I'll just add a random vector to the actions. And to make them more varied, I should make that vector dependent on the input to the head. That can be done by simply multiplying it by a random matrix. That should do wonders for making semi-realistic players.

The real challenge then will be to discover who can exploit the best.

I'll also try this with recurrent nets. Even bad players adapt to an extent so that might be even better.

2:05pm. No, the learning rate is too high. I need to lower it.

After that I'll compare it with a two layer `mi_prong_alt`.

2:10pm.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 211 and 789 out of 1000.
The time was 00:00:15.0344871 for: iteration 0
Starting timing for: iteration 1
Winrate is 234 and 766 out of 1000.
The time was 00:00:23.2258506 for: iteration 1
Starting timing for: iteration 2
Winrate is 571 and 429 out of 1000.
The time was 00:00:35.1860057 for: iteration 2
Starting timing for: iteration 3
Winrate is 704 and 296 out of 1000.
The time was 00:00:36.9200506 for: iteration 3
Starting timing for: iteration 4
Winrate is 699 and 301 out of 1000.
The time was 00:00:35.3638805 for: iteration 4
Starting timing for: iteration 5
Winrate is 714 and 286 out of 1000.
The time was 00:00:31.3819817 for: iteration 5
Starting timing for: iteration 6
Winrate is 743 and 257 out of 1000.
The time was 00:00:29.9635441 for: iteration 6
Starting timing for: iteration 7
Winrate is 721 and 279 out of 1000.
The time was 00:00:31.3875519 for: iteration 7
Starting timing for: iteration 8
```

This is just such a strange learning cure. It spends a bunch of time doing nothing and then explodes upward. This is not like feedforward nets. The feedforward nets do it more systematically. I think RNNs as well, but I do not think I've yet to try 2 layer nets. The reason why I thought that -10 was fine is because it only got 210 on the first iteration with it as well. I wonder if I am initializing these poorly? I wrote about how I should probably use larger initial weigths for the poker game.

2:15pm. When the tests are done I really should put some effort in making a SNAIL like architecture.

Basically standard RNNs are hopeless for my purpose. I really only the have the option of DP or a Wavenet like architecture.

Given how poorly a two layer plastic RNN works on the Binary Pattern test, I am really tempted to consider alternative networks that could do it. Thought LSTMs and standard recurrent nets are powerful they cannot really learn across many timesteps. This is something WaveNet style architectures have solved. They are a really powerful alternative and something I would reach for if DP did not get here. But I should seriously consider them anyway as they are both memory efficient than current plastic RNNs and are more fit for use on the GPU.

But RNNs are definitely more powerful and are more computationally efficient. And I have PRONG on my side.

Now that I've lowered the learning rate the net is exceptionally stable. I think I should disregard the results from the one layer net. -10 was too high despite the initial appearance. The way plastic RNNs learn is different from feedforward ones. It might also be the intialization.

2:35pm. I think the net passes after all. The results I am getting in the later stages now are trully good. I can make do with this. In order to protect against bad init, I think I should implement Oja's rule. That video I've stopped halfway did mention that it normalizes the Hebbian traces towards 1. That could be super useful. I definitely need that.

2:40pm. Ok, I think this should suffice as a plan. Let me finish off the Hebbian learning video. I've written enough."