Yeah, better init makes it learn on part with `mi_prong_alt`.
Enough of this. Either way, the learning abilities of the RNNs are remarkable and they are stable. Everything is going according to plan.

If I decide to go with platic RNNs I will have a bunch of clean up work left for me. I'll have to make a custom kernel for the Hebbian update. And I need to make a activation function that allows automatic updating of covariance matrices. Which I will need for the generalized mi activation. That one will be on the TODO list regardless as I will use the MI activation in the temporal feedforward NN (TFNN). This is the sort of thing is what Spiral was made for. The guys using PyTorch won't be doing stuff like this.

Today was such a lazy slow day. But it is fine to have days like these every once in a while. I'll really let my power loose tomorrow.

Edit: Forgot to commit this 3h ago.

As I wrote the mail to the author I had a change of heart regarding plasticity. Yeah, sure that it is memory inefficient, but that is only the case if I use it like I was doing now.

My sense is definitely that something is amiss in how I am doing things currently.

Instead of following the paper directly, I should make my own architecture with plastic memory cells. That will make the optimization problem a lot easier.

Also temporal architectures or to better put it SNAIL in particular is a lot more complicated than I remember it being. Yeah, sure I remembered the most salient points, but I forgot just how complex the implementation was. It is not wonder that I forgot the method itself up to now.

I need to focus on recurrent nets for the time being.

Small memory cells. What an amazing idea. It is amazing that it did not occur to me even once during the night.

Also on second thought I am not buying the Maze results. I think it is taking 100s of 1000s of episodes simply because they are having variance issues. If I gave it a try I could probably get it to run a lot better with PRONG.

Well, let me call it a day here. Somehow I ended up putting quite a lot of overtime today.

Plasticity is fresh and I have nobody to point the way to me with regards to how to use it. I think that tomorrow I will take a look at that reference on neuromodulation in the DP paper.

I might have nobody to point the way, but I should at least be able to realize that if I want to grasp my goal, I should take a step forward and do it myself. That goes for both plasticity and in general..

I cannot abandom my approach just because of this much difficulty. This cannot possibly be the full power of differentiable plasticity. I should be able to draw out much more than this.