I've been chilling for the past 45m ever since I got up as usual. I am still thinking about what I want to do.
I think today I will stop programming for a bit and do some study of neuromodulation. I'll also learn more about Hebbian learning.

I simply understand far too little about how plasticity should really be used. Since I've decided that I do not want temporal connections that makes things a lot simpler regarding what I want to do from here on out.

Ultimately, all these choices are there to constrain my search. With the decision I made, that makes it a lot easier to decide of how I want to move from here. Deciding not to look into temporal architectures means that my only choice is plasticity.

There is no better way of starting the journey than this particular architecture.

The network won't be a 10 layer monster to start things off, but a simple 1,2 or 3 layer plastic net. Hell, I should even consider doing it on the CPU if that is the case.

I am just joking. But one should not start of with 10 layers. 1 layer is what makes the most sense. This is much like when learning programming one starts with small programs and builds up from there.

Tonight was actually the first time I seriously thought about adding checkpointing. And yesterday was the first time I sensed that is time to upgrade the game.

Now that it has been carved in stone what I will use I need to decide on my priorities:

1) I need to do more research on biologically inspired learning. The Binary Pattern can barely be considered solved right now. It should be possible to do much better.

I need better architectures that take true advantage of plasticity. The one I have now is entry level.

2) Extend the game. I am going to need this in order to properly test whether plasticity is enabling actual metalearning. Not just that, but this is an inevitable step anyway as I work my way to NL Holdem.

3) Make hand made implementations of the Hebbian update, and fold the covariance matrix updates into broadcasting_activation.

After that the paths branch:

- Experiment on different architectures. Hopefully the research I will do now will give me some ideas. The Hebbian learning video already mentioned suppresive connections that the current plasticity framework lacks.

- I should probably bring in Oja's rule at some point.

- Add model checkpointing. The training times will get only bigger from here. But Spiral is well suited for implementing this sort of thing. It won't be a problem.

- I should let `n` be a trainable parameter. The instability I got was because of incorrectly done PRONG. This I will do when making the hand made Hebbian update. I should save it for the Oja's rule.

- I should forget Rudder for a month or two. I do not have time to fit it in between everything else. My schedule for the next month will be pretty stuffed. It is a priority, but not an immediate one.

I'll keep it out of mind and get to it after I get everything else in order. I do not want to needlessly obligate myself with it and overburden myself as a result.

Rudder is an advanced method.

- The next two month I should dedicate to getting the basics of plasticity down much like past few months were dedicated to natural gradients and RL in general.

There are all sorts of things coming out every day and I cannot keep up with it all. Most of it is filler anyway, so I do not have to worry about it. If I miss something important, I'll find out about it anyway in due time.

- I still haven't come close to testing out Zap in standard layers. Like Rudder, this will be another thing I will leave for later. I'll forget it even exists until then.

---

I've been writing the above for more than hour now which is usual for large rants. It might seem like a waste of time, but they are necessary in order to gather my focus. Let me sharpen it more by outlining the following.

- Simplicity is the key to everything. Simplicity is what will allow me to succeed. Good programming is simplifying things to their limit. The more a programmer is capable of minimizing the cognitive burden the better he is. The same applies to the algorithm used. Hebbian learning will undobtedly play a role as a localized learning rule just as it does in biological systems.

And PRONG is a localized second order optimizer. These good local rules are incredibly valuable and must be collected at all cost. They are the only kinds of rules that scale.

I cannot follow the same path as the OpenAI guys it seems. What I really want to do and what I wanted to do all this time is master the basics.

- From here on out and until the end of the year, mastering the basics is all that I will be doing.

By the time I enter the arena for real, the architecture I will be using will seem like absolutely nothing as it will be so simple. It will be simple to use and simple to explain.

- From here on out and until the end of the year, the progress I will be making will finally be realistic. It has been this way since April. For every month that has passed since then, my understanding of RL has improved by a significant amount.

Right now it feels like early 2016 all over again, except all the pieces are in place and I can finally make the steps that I need.

I should yield to that feeling. I want to pretend that the horrifying expenditure of time of the last 3.5 years did not happen and go on as a starry eyed beginner once again. The last 3.5 years do not matter. But the following ones certainly will.

I spent my 20s chasing an impossible goal and relying on my own power. I will erase the shame of that. As I have ever since I decided to become a programmer, I will dedicate my life to chaining simple goals together, this time from the bottom rather than the top. The Spiral language and differentiable programming will give me the strength to realize those ideals.

And I will never rely on myself or others. Only the power of machines.