Today as per plan, I will be doing studying. I want to follow up on some references and make a tour of EPANN papers just to see what kinds of different architectures there are.
https://sc2ai.net/Tournament.php

I want to join this to grab that Titan X, but I guess things like these are for the future. Now comes research, hardcore research.

One day I'll be able to take my pick of these for fun. But now I need to build up the capability to win them.