``` RNN.mi_hebb'_prong n 128, RNN.mi_hebb'_prong n 128, RNN.mi_hebb'_prong n 128 ```
With a single layer it takes 3s to compile. With this, it takes 4.65s. With only feedforward layers it at 1.8s. So there are some compilation issues. I will squish them later. It is most likely doing too much inlining. I've already gone through this once with LSTMs. This time I hopefully won't have to consider rewriting the language in Ocaml or Racket.

Well, I've investigated this. Let me have some fun. Are the Nvidia benchmarks out yet? It has been a few weeks so they should be.