"8:05am. I am up. I figured out what Hebbian learning is finally doing. Or rather what it is supposed to be doing. I was actually rather confused as to what it is supposed to be doing for a while now. This is because the DP formulation actually makes very little sense.
Let me start from the top.

Hebbian learning on its own does PCA. Without inhibitory connections all the weights would converge to the same principal component. Oja's rule causes weight normalization. Interestingly, autoencoders themselves do PCA, but Hebbian learning is simpler.

What modulation does is splits the input into segments allowing the Hebbian learning to learn separate representations. So it is absolutely vital to its functioning - all the weights learning the same thing is just not useful.

If I train the modulators with backprop and the regular weights with Hebbian learning that would allow for the ideal combination of supervised and unsupervised learning. It all makes sense now.

Note that the `W` and `A` in `W + A * H` do play a modulatory role somewhat, but it a very complicated way of doing it and I am not surprised that the net just dies with two layers on the Binary Pattern test. It is obvious to me that what I introduced yesterday is a far superior way of doing it. It directly does exactly what it has to.

It should work, but it always possible that Miconi tried this and discarded it. But if it works, this is the kind of idea that will upend the way deep learning is done.

Well, one could replace Hebbian learning with regular autoencoder updates and it would still work, though it would be more complicated for the same effect so there is not much point to it. The closest thing I saw to this in terms of online learning thing would be dynamic evaluation. I think this is better though.

I've been thinking about this for 30m so let me do some more. I am thinking about whether I should use multiplicative integration everywhere and I am leaning towards yes. With different modulators and different inputs the two weight matrices in the RNN should be learning different things despite having the same outputs. It is no problem.

I am good to go here. For the Binary Pattern test, I am going to need an RNN. I do not think a feedforward net will be able to do this.

8:40am. It is good to do things slowly so all the pieces settle into their proper place. Now that I know that I am doing semi supervised learning, I feel like doing some studying on that. I am not up to date with it. I feel like reading the Ladder network paper again and moving from there.

Once I am done with work on differentiable plasticity, it might be a good idea to do some fishing for a Kaggle contest that has a large amount of unlabeled data. The architecture I am using would give me an edge over the competition.

One thing I feel right now is watching the videos on RBMs by Hinton. I started this by studying Hopfield nets and I want to finish it. I know that I already did this years ago, but I did not understand it back then so I haven't internalized it properly.

Let me also say one other thing. Apart from the Binary Pattern test, I am quite sure based on the training curves that the plastic part of the Hebbian net is not getting any workout on poker and the char RNN. With my current scheme I won't have to worry about that.

8:55am. Let me chill a little as I usually do in the morning. Then come studies and then I will move to testing the thing I did yesterday."